requirements.txt
webdriver-manager 
selenium
json
urllib
beautifulsoup4
argparse
pdf2image
python-dotenv
google-generativeai